WIP Added RoleFitness.h
